stash: restore the index by default when breaking changes are enabled
This footgun is described in the previous commit to
Documentation/BreakingChanges.adoc; without --index, stash apply and
stash pop confusingly drop stashed index changes.

When compiling with breaking changes, instead restore the index unless
asked not to. Adjust both argument parsing and handling, as well as help
synopses. Don't forget to adjust the error message when application
fails to point to --no-index instead.

Signed-off-by: D. Ben Knoble <ben.knoble+github@gmail.com>