t3903: adjust stash test to account for --[no-]index with breaking changes
A few tests check the results of the index after applying a stash; with
breaking changes from previous commits that automatically restore the
stashed index, the expected values are wrong.

A few of the relevant tests check the restoration of <pathspec>s; with
the aforementioned breaking changes, things get more interesting. In
particular, if we "git stash push -- foo" but have "bar" in the index,
then when applying the stash we get a conflict: "bar" was not removed
from the index by the stash, but it was included in the recorded index
in the stash. In those cases, apply the stash with "--no-index" (which
would be the required user behavior).

Signed-off-by: D. Ben Knoble <ben.knoble+github@gmail.com>