editor: use standard strvec API to receive environment for external editors
Going back to the introduction of the env parameter for the editor in
8babab95af (builtin-commit.c: export GIT_INDEX_FILE for launch_editor as
well., 2007-11-26), we pass a constant array of strings: as the
surrounding APIs evolved to use strvecs, the editor code did not.

There is only one caller of all 3 editor APIs that does not pass a NULL
environment (the same caller for which this parameter was added), and
it already has a strvec available to use.

Signed-off-by: D. Ben Knoble <ben.knoble+github@gmail.com>
Helped-by: Johannes Schindelin <johannes.schindelin@gmx.de>